pyproject.toml
[project]
name = "speed-mcp-server"
version = "0.1.0"
description = "MCP server that holds 0x, Alchemy, Squid, and OpenSea API keys so Speed-CLI and agents can use them without local config. Supports stdio (Cursor) and Streamable HTTP (remote); optional encrypted delivery via client RSA public key. Wallet PRIVATE_KEY stays local in ~/.speed only."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "cryptography>=43.0.0",
  "httpx>=0.27.0",
  "mcp>=1.23.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
speed-mcp = "speed_mcp_server.main:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["speed_mcp_server*"]
